requirements.txt
flask
nltk
tiktoken
pillow
numpy
librosa
matplotlib
soundfile
scipy
werkzeug